Change factory connectiond etails from provider to consumer
